crm_backend/requirements.txt
Django>=4.2,<5.0
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-cors-headers 